feat: Introduce Flake for development and builds
This PR introduces a Nix flake, allowing for InfiniSim to be built as a
Flake, including a FHS development environment.

We also introduce `flake-compat`, allowing for non-Flake Nix mahcines to
use the project as-is, both for building (`default.nix`), and
development (`shell.nix`).

Additionally, we introduce `.envrc`, meaning that with `direnv`, the Nix
Flake is activated automatically.

It should be noted that we require
https://github.com/InfiniTimeOrg/InfiniTime/pull/2121 to be merged
into mainline before this PR will work. It will currently fail - this is
expected at this stage. A lockfile needs to be generated after the above
PR is merged - this PR here should NOT be merged until that is
completed.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>