Conditionally pass Controllers::FS to AlarmController to support save to file
Support for InfiniTimeOrg/InfiniTime#1367 which is a rebase/improvement
of InfiniTimeOrg/InfiniTime#1333 by @ght

The PR adds saving and restoring alarm state to the SPI NOR flash. So
the AlarmController needs the FileSystem-controller to work.

To support both the current `main` branch and the "save-alarm-to-file"
PR pass the Filesystem controller conditionally, only if
`Controller::FS` is mentioned in the `AlarmController.h` file