Add a workaround for error "Status 500: unable to find user ubuntu: no matching entries in passwd file" when build dev container in CLion
Clion tries to read the local /etc/password when building to change the UID of ther user in the container to the one from outside. If your /etc/passwd does not have this entry, it will fail to build.
Comment this in if you do want to use the dev container CLion (at least CLion 2024.3.4, Build #JBC-243.25659.42. It is reported to Jetbrains).