Add Dockerfile and devcontainer.json (#174)
This allows both interactive use in IDEs/editors with dev container support as well as building inside the dev container from the command line.
The changes are based on https://github.com/kon-foo/InfiniSim/tree/docker-builder-image.

The original PR was extended with the following features:
- ccache and ninja are installed for optional faster builds. Using ninja instead of make speeds up the build from ~ 28 s to 17 s on my machine with 16 threads.
- git and sudo are added for interactive use of the devcontainer
- Store bash history in volume to preserve it across rebuilds
- The non-root user infinitime with sudo password "it" was added and is used to avoid producing build artifacts owned by root
- Split args into CMake generation time and CMake build time args
- Allow passing the build directory in docker run

- Change the container user to ubuntu:ubuntu since that one already exists anyway. There were also permission issues stemming from not using ubuntu since the external user would likely have UID 1000 (at least on single user systems). That would lead to external files mapped into the container being owned by ubuntu, not infinitime, which has UID 1001
- Split installation into three sections to increase layer granularity for caching
- Remove node install script after usage to reduce container size
- Add gdb to allow debugging inside of the dev container
- Set name of dev container bash history volume to infinisim-bashhistory


Add a workaround for error "Status 500: unable to find user ubuntu: no matching entries in passwd file" when build dev container in CLion

Clion tries to read the local /etc/password when building to change the UID of ther user in the container to the one from outside. If your /etc/passwd does not have this entry, it will fail to build.
Comment this in if you do want to use the dev container CLion (at least CLion 2024.3.4, Build #JBC-243.25659.42. It is reported to Jetbrains).

Co-authored-by: kon-foo <25391223+kon-foo@users.noreply.github.com>
Co-authored-by: NeroBurner <pyro4hell@gmail.com>