Make dev container compatible with podman (#178)
Following the issue described in https://github.com/InfiniTimeOrg/InfiniSim/pull/174#discussion_r2150883468.
This prevents permission issues when using the dev container with `podman`, which is running rootless.
Usage with `docker` requires explicitly passing `UID` and `GID` of the local user, which is added to the README.

Add podman command to `README.md`

Co-authored-by: NeroBurner <pyro4hell@gmail.com>

---------

Co-authored-by: NeroBurner <pyro4hell@gmail.com>